# 10x speedup for ZSTD_isRLE()